Net connect jlong

+++ src/java.base/windows/native/libnio/ch/Net.c
@@ -43,7 +43,7 @@
 #include "sun_nio_ch_Net.h"
 #include "sun_nio_ch_PollArrayWrapper.h"
 
-#include<ut_jcl_nio.h>
+#include "ut_jcl_nio.h"
 
 /**
  * Definitions to allow for building with older SDK include files.
@@ -249,10 +249,10 @@ Java_sun_nio_ch_Net_connect0(JNIEnv *env, jclass clazz, jboolean preferIPv6, job
 
     if (AF_INET == sa.sa4.sin_family) {
         char buf[INET_ADDRSTRLEN];
-        Trc_nio_ch_Net_connect4(s, inet_ntop(AF_INET, &sa.sa4.sin_addr, buf, sizeof(buf)), port, sa_len);
+        Trc_nio_ch_Net_connect4((jlong)s, inet_ntop(AF_INET, &sa.sa4.sin_addr, buf, sizeof(buf)), port, sa_len);
     } else if (AF_INET6 == sa.sa6.sin6_family) {
         char buf[INET6_ADDRSTRLEN];
-        Trc_nio_ch_Net_connect6(s, inet_ntop(AF_INET6, &sa.sa6.sin6_addr, buf, sizeof(buf)), port, ntohs(sa.sa6.sin6_scope_id), sa_len);
+        Trc_nio_ch_Net_connect6((jlong)s, inet_ntop(AF_INET6, &sa.sa6.sin6_addr, buf, sizeof(buf)), port, ntohs(sa.sa6.sin6_scope_id), sa_len);
     }
 
     rv = connect(s, &sa.sa, sa_len);
